requirements.txt
bleach
boto3
Commitizen
commitizen-emoji
django
django-amazon-ses
django-cors-headers
django-environ
django-health-check
django-jazzmin
django-private-storage
django-ses
django-simple-history
django-split-settings
django-storages
djangorestframework
djangorestframework_simplejwt
djoser
drf-yasg
drf-yasg[validation]
drfpasswordless
factory-boy
mysqlclient==2.0.3
Pillow
pre-commit
pyjwt
pytest
pytest-cov
pytest-django
stripe
django-ebhealthcheck
django-q
django-import-export
